+++ user/app/bsp_vlps.h
@@ -1,0 +1,17 @@
+
+/*
+ *
+ * Change Logs:
+ * Date           Author       Notes
+ * 2019-6-04     suozhang      the first version
+ *
+ */
+
+
+#ifndef BSP_VLPS_H
+#define BSP_VLPS_H
+
+void vlps_init( void );
+
+#endif /* BSP_VLPS_H */
+
